Switch from JSON to Persistence API
Since in-package JSON cannot be modified, I figured it's better
to make a switch to profile configuration with Persistence API.
Gson was removed as a dependency, may make a return if I decide
that JSON is the best option for online profile data transfer.

Additional minor change:
Project structure was rearranged a little to better accomodate
a few classes that were previously in ui/element submodule.