Adjust build.gradle for jar packaging
Deprecated use of javafxplugin and jlink.
Linked JavaFX dependencies and adjusted jar task to include them.
Delegated main entrypoint to new Run.java class (workaround)

Kudos to JavaFX subreddit and @Col-E for the guides!

Also, while at it, changed the name of the package.
It's now aligned with the actual name of the project.