Implement server communication
This is a pretty big change - the online side of things
has just gotten its scaffold! Not all functionality is implemented,
but the general client-server communication is DONE.

You can play around in an online lobby now. Not much is there,
so far only joining, leaving, (un)readying and data exchange,
and by default it works on LAN only - no forwarding.

There's a bunch of actions next in line to be implemented like chat
and in-game signaling (playing/drawing cards, skipping etc).

A lot of the code was refactored to support online functionality,
and a lot more will have to be. It takes quite a lot of time.
So far it was mostly profiles that were split into local
(personal profile) and export (derived, used in games).
Lobby was also adjusted into local and online separately.
Game superclass (and probably subclasses) will be next.
Maybe other entity classes will need it as well.

This is all highly WIP, so bugs are likely. API changes will be
very dynamic over the course of this stage of development.

I worked with networking already, extended server-side functionality,
but implementing such module from scratch is pretty neat tbh.
I feel a lot of satisfaction from fairly small things.
Seeing the players appear and disappear on the player lists,
ready mark change from red to green (and green to red)
on server and other clients. It's not much, but... it's a start.

The communication side is ready. Off to connecting it to logic :)